to ensure that it goes through all of the vector files.
hmm, lemme see… if the doop size is = i then it might actually allow it
to go one over… which might be a seg fault… should be less than.
Shouldn’t it? Yeah, pretty sure it should. If we start at one, we are
ignoring 0 since that is the execution file doop size will give us an
int, not taking into account the 0 index… so it should be less than.
Changing back to <.